refactor(api): own catalog types in Zod and split client barrel
Drop hand-written CategoryTree duplicates and OrganizationData naming so the catalog payload has a single Zod source of truth, with fetch helpers in client.ts and index as exports only.
